requirements.txt
mcp>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.1
email-validator>=2.1.1
psycopg[binary]>=3.2.0